requirements.txt
asgiref==3.9.2
charset-normalizer==3.4.3
cssselect2==0.8.0
Django==5.2.7
django-filer==3.3.2
django-polymorphic==4.1.0
easy-thumbnails==2.10.1
freetype-py==2.5.1
lxml==6.0.2
pillow==11.3.0
pycairo==1.28.0
reportlab==4.4.4
rlPyCairo==0.4.0
sqlparse==0.5.3
svglib==1.6.0
tinycss2==1.4.0
webencodings==0.5.1
